third_party/blink/web_tests/fast/forms/calendar-picker/date-open-picker-with-f4-key.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/js-test.js"></script>
<script src="resources/calendar-picker-common.js"></script>

<p id="description"></p>
<div id="console"></div>
<input type=date id=test>
<script>
description("Tests if pressing F4 opens the calendar picker.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('internals.pagePopupWindow');

debug('Press F4 key.');
sendKey(document.getElementById('test'), 'F4');
shouldBeNonNull('internals.pagePopupWindow');

function sendKey(input, keyName) {
    input.focus();
    eventSender.keyDown(keyName);
}
</script>
